Pull fee into its own command
