+++ cmd/soroban-cli/src/commands/tx/fetch/fee.rs
@@ -1,0 +1,211 @@
+use crate::{
+    rpc,
+    commands::global,
+    config::network,
+    xdr::{self, Limits, WriteXdr, Hash, TransactionMeta, SorobanTransactionMetaExt},
+};
+use clap::{command, Parser};
+use prettytable::{format::{FormatBuilder, LinePosition, LineSeparator}, Table, Row, Cell};
+use serde::{Deserialize, Serialize};
+use soroban_rpc::GetTransactionResponse;
+use super::args;
+
+#[derive(Parser, Debug, Clone)]
+#[group(skip)]
+pub struct Cmd {
+    #[command(flatten)]
+    pub(crate) args: FeeArgs,
+}
+
+#[derive(Debug, Clone, clap::Args)]
+pub struct FeeArgs {
+    /// Transaction hash to fetch
+    #[arg(long)]
+    pub hash: Hash,
+
+    #[command(flatten)]
+    pub network: network::Args,
+
+    /// Output format for fee command
+    #[arg(long, default_value = "json")]
+    pub output: FeeOutputFormat,
+}
+
+impl FeeArgs {
+    pub async fn fetch_transaction(
+        &self,
+        global_args: &global::Args,
+    ) -> Result<GetTransactionResponse, Error> {
+        let network = self.network.get(&global_args.locator)?;
+        let client = network.rpc_client()?;
+        let tx_hash = self.hash.clone();
+        let tx = client.get_transaction(&tx_hash).await?;
+        match tx.status.clone() {
+            val if val == "NOT_FOUND".to_string() => {
+                if let Some(n) = &self.network.network {
+                    return Err(Error::NotFound {
+                        tx_hash,
+                        network: n.to_string(),
+                    });
+                }
+            }
+            _ => {}
+        }
+        Ok(tx)
+    }
+}
+
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, clap::ValueEnum, Default)]
+pub enum FeeOutputFormat {
+    /// JSON output of the ledger entry with parsed XDRs (one line, not formatted)
+    #[default]
+    Json,
+    /// Formatted (multiline) JSON output of the ledger entry with parsed XDRs
+    JsonFormatted,
+    /// Original RPC output (containing XDRs)
+    Xdr,
+    /// Formatted in a table comparing fee types
+    Table,
+
+}
+
+#[derive(thiserror::Error, Debug)]
+pub enum Error {
+     #[error(transparent)]
+    Network(#[from] network::Error),
+    #[error(transparent)]
+    Serde(#[from] serde_json::Error),
+    #[error(transparent)]
+    Xdr(#[from] xdr::Error),
+    #[error(transparent)]
+    Args(#[from] args::Error),
+    #[error("{message}")]
+    NotSupported { message: String },
+    #[error("transaction {tx_hash} not found on {network} network")]
+    NotFound { tx_hash: Hash, network: String },
+    #[error(transparent)]
+    Rpc(#[from] rpc::Error),
+}
+
+impl Cmd {
+    pub async fn run(&self, global_args: &global::Args) -> Result<(), Error> {
+        let resp = self.args.fetch_transaction(global_args).await?;
+        let tx_result = resp.clone().result.unwrap();
+        let tx_meta = resp.clone().result_meta.unwrap();
+        let fee = tx_result.fee_charged;
+        let (non_refundable_resource_fee, refundable_resource_fee) = match tx_meta.clone() {
+           TransactionMeta::V0(_) => {
+                return Err(Error::NotSupported { message: "TransactionMeta::V0 not supported".to_string() });
+            },
+            TransactionMeta::V1(_) => {
+                return Err(Error::NotSupported { message: "TransactionMeta::V1 not supported".to_string() });
+            },
+            TransactionMeta::V2(_) => {
+                return Err(Error::NotSupported { message: "TransactionMeta::V2 not supported".to_string() });
+            },
+            TransactionMeta::V3(meta) => {
+                if let Some(soroban_meta) = meta.soroban_meta {
+                    match soroban_meta.ext {
+                        SorobanTransactionMetaExt::V0 => {
+                            return Err(Error::NotSupported { message: "SorobanTransactionMetaExt::V0 not supported".to_string() })
+                        },
+                        SorobanTransactionMetaExt::V1(v1) => {
+                            (v1.total_non_refundable_resource_fee_charged, v1.total_refundable_resource_fee_charged)
+                        },
+                    }
+                } else {
+                    return Err(Error::NotSupported { message: "cannot get fee when soroban_meta is None".to_string()})
+                }
+            },
+        };
+
+        let fee_table = FeeTable{ fee, non_refundable_resource_fee, refundable_resource_fee };
+
+        match self.args.output {
+            FeeOutputFormat::Json => {
+                        println!("{}", serde_json::to_string(&fee_table)?);
+                    }
+            FeeOutputFormat::Xdr => {
+                        return Err(Error::NotSupported { message: "xdr output is not available for the fee subcommand".to_string() })
+                    }
+            FeeOutputFormat::JsonFormatted => {
+                        println!("{}", serde_json::to_string_pretty(&fee_table)?);
+                    }
+            FeeOutputFormat::Table => fee_table.print(),
+        }
+
+        Ok(())
+    }
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct FeeTable {
+    fee: i64,
+    non_refundable_resource_fee: i64,
+    refundable_resource_fee: i64,
+}
+
+impl FeeTable {
+    fn inclusion_fee(&self) -> i64 {
+        self.fee - self.resource_fee()
+    }
+    
+    fn resource_fee(&self) -> i64 {
+        self.non_refundable_resource_fee + self.refundable_resource_fee
+    }
+
+    fn print(&self) {
+        let table_format = FormatBuilder::new()
+                             .column_separator('│')
+                             .borders('│')
+                             .separators(&[LinePosition::Top],
+                                         LineSeparator::new('─',
+                                                            '─',
+                                                            '┌',
+                                                            '┐'))
+                             .separators(&[LinePosition::Intern],
+                                         LineSeparator::new('─',
+                                                            '─',
+                                                            '├',
+                                                            '┤'))
+                             .separators(&[LinePosition::Bottom],
+                                         LineSeparator::new('─',
+                                                            '─',
+                                                            '└',
+                                                            '┘'))
+                             .padding(1, 1)
+                             .build();
+
+        let mut table = Table::new();
+
+        // Optional: customize borders
+        // table.set_format(*format::consts::FORMAT_BOX_CHARS);
+        table.set_format(table_format);
+
+        // First row: single wide cell (horizontally spans 2 columns)
+        table.add_row(Row::new(vec![
+            Cell::new(&format!("tx.fee: {}", self.fee))
+                .style_spec("b")
+                .with_hspan(3),
+        ]));
+
+        // Second row: two separate cells
+        table.add_row(Row::new(vec![
+            Cell::new(&format!("tx.v1.sorobanData.resourceFee: {}", self.resource_fee()))
+                .style_spec("FY")
+                .with_hspan(2),
+            Cell::new(&format!("inclusion fee: {}", self.inclusion_fee())),
+        ]));
+
+        table.add_row(Row::new(vec![
+            Cell::new(&format!("fixed resource fee: {}", self.non_refundable_resource_fee))
+                .style_spec("FY"),
+            Cell::new(&format!("refundable resource fee: {}", self.refundable_resource_fee))
+                .style_spec("FY"),
+            Cell::new(&format!("inclusion fee: {}", self.inclusion_fee())),
+        ]));
+
+        table.printstd();
+    }
+}
