+++ cmd/soroban-cli/src/commands/tx/fetch/mod.rs
@@ -1,7 +1,7 @@
 use clap::{command, Subcommand};
 use std::fmt::Debug;
 use prettytable::{format::{self, FormatBuilder, LinePosition, LineSeparator}, Cell, Row, Table};
-
+use serde::{Deserialize, Serialize};
 use crate::{
     commands::global,
     config::network,
@@ -16,6 +16,7 @@ mod args;
 mod envelope;
 mod meta;
 mod result;
+mod fee;
 
 #[derive(Debug, clap::Args)]
 #[command(args_conflicts_with_subcommands = true)]
@@ -33,6 +34,8 @@ pub enum FetchCommands {
     Result(result::Cmd),
     /// Fetch the transaction meta
     Meta(meta::Cmd),
+    // Fetch the transaction fee information
+    Fee(fee::Cmd),
     /// Fetch the transaction envelope
     #[command(hide = true)]
     Envelope(envelope::Cmd),
@@ -50,9 +53,6 @@ struct DefaultArgs {
     /// Format of the output
     #[arg(long, default_value = "json")]
     pub output: Option<args::OutputFormat>,
-
-    #[arg(long)]
-    pub fee_only: bool
 }
 
 #[derive(thiserror::Error, Debug)]
@@ -65,20 +65,17 @@ pub enum Error {
     Meta(#[from] meta::Error),
     #[error(transparent)]
     Envelope(#[from] envelope::Error),
-    #[error("{message}")]
-    NotSupported { message: String },
+    #[error(transparent)]
+    NotSupported(#[from] fee::Error),
 }
 
 impl Cmd {
     pub async fn run(&self, global_args: &global::Args) -> Result<(), Error> {
-        if self.default.fee_only {
-            return self.fee_only(global_args).await
-        }
-
         match &self.subcommand {
             Some(FetchCommands::Result(cmd)) => cmd.run(global_args).await?,
             Some(FetchCommands::Meta(cmd)) => cmd.run(global_args).await?,
             Some(FetchCommands::Envelope(cmd)) => cmd.run(global_args).await?,
+            Some(FetchCommands::Fee(cmd)) => cmd.run(global_args).await?,
             None => {
                 envelope::Cmd {
                     args: args::Args {
@@ -96,119 +93,5 @@ impl Cmd {
             }
         }
         Ok(())
-    }
-
-    async fn fee_only(&self, global_args: &global::Args) -> Result<(), Error> {
-        let args =  args::Args {
-            hash: self.default.hash.clone().unwrap(),
-            network: self.default.network.clone().unwrap(),
-            output: self.default.output.unwrap(),
-        };
-
-        let resp = args.fetch_transaction(global_args).await?;
-        let tx_result = resp.clone().result.unwrap();
-        let tx_meta = resp.clone().result_meta.unwrap();
-        let fee = tx_result.fee_charged;
-        let (non_refundable_resource_fee, refundable_resource_fee) = match tx_meta.clone() {
-           TransactionMeta::V0(_) => {
-                return Err(Error::NotSupported { message: "TransactionMeta::V0 not supported".to_string() });
-            },
-            TransactionMeta::V1(_) => {
-                return Err(Error::NotSupported { message: "TransactionMeta::V1 not supported".to_string() });
-            },
-            TransactionMeta::V2(_) => {
-                return Err(Error::NotSupported { message: "TransactionMeta::V2 not supported".to_string() });
-            },
-            TransactionMeta::V3(meta) => {
-                if let Some(soroban_meta) = meta.soroban_meta {
-                    match soroban_meta.ext {
-                        SorobanTransactionMetaExt::V0 => {
-                            return Err(Error::NotSupported { message: "SorobanTransactionMetaExt::V0 not supported".to_string() })
-                        },
-                        SorobanTransactionMetaExt::V1(v1) => {
-                            (v1.total_non_refundable_resource_fee_charged, v1.total_refundable_resource_fee_charged)
-                        },
-                    }
-                } else {
-                    return Err(Error::NotSupported { message: "cannot get fee when soroban_meta is None".to_string()})
-                }
-            },
-        };
-
-        FeeTable{ fee, non_refundable_resource_fee, refundable_resource_fee }.print();
-
-        Ok(())
-    }
-
-}
-
-struct FeeTable {
-    fee: i64,
-    non_refundable_resource_fee: i64,
-    refundable_resource_fee: i64,
-}
-
-
-impl FeeTable {
-    fn inclusion_fee(&self) -> i64 {
-        self.fee - self.resource_fee()
-    }
-    
-    fn resource_fee(&self) -> i64 {
-        self.non_refundable_resource_fee + self.refundable_resource_fee
-    }
-
-    fn print(&self) {
-        let table_format = FormatBuilder::new()
-                             .column_separator('│')
-                             .borders('│')
-                             .separators(&[LinePosition::Top],
-                                         LineSeparator::new('─',
-                                                            '─',
-                                                            '┌',
-                                                            '┐'))
-                             .separators(&[LinePosition::Intern],
-                                         LineSeparator::new('─',
-                                                            '─',
-                                                            '├',
-                                                            '┤'))
-                             .separators(&[LinePosition::Bottom],
-                                         LineSeparator::new('─',
-                                                            '─',
-                                                            '└',
-                                                            '┘'))
-                             .padding(1, 1)
-                             .build();
-
-        let mut table = Table::new();
-
-        // Optional: customize borders
-        // table.set_format(*format::consts::FORMAT_BOX_CHARS);
-        table.set_format(table_format);
-
-        // First row: single wide cell (horizontally spans 2 columns)
-        table.add_row(Row::new(vec![
-            Cell::new(&format!("tx.fee: {}", self.fee))
-                .style_spec("b")
-                .with_hspan(3),
-        ]));
-
-        // Second row: two separate cells
-        table.add_row(Row::new(vec![
-            Cell::new(&format!("tx.v1.sorobanData.resourceFee: {}", self.resource_fee()))
-                .style_spec("FY")
-                .with_hspan(2),
-            Cell::new(&format!("inclusion fee: {}", self.inclusion_fee())),
-        ]));
-
-        table.add_row(Row::new(vec![
-            Cell::new(&format!("fixed resource fee: {}", self.non_refundable_resource_fee))
-                .style_spec("FY"),
-            Cell::new(&format!("refundable resource fee: {}", self.refundable_resource_fee))
-                .style_spec("FY"),
-            Cell::new(&format!("inclusion fee: {}", self.inclusion_fee())),
-        ]));
-
-        table.printstd();
     }
 }
